Int32 to Int64 migration.

+++ trunk/src/com/geoscope/GeoEye/Space/Functionality/TTypeFunctionality.java
@@ -39,7 +39,7 @@ public class TTypeFunctionality extends TFunctionality {
 	public void DestroyInstance(long idComponent) throws Exception {
 		String URL1 = Server.Address;
 		//. add command path
-		URL1 = "http://"+URL1+"/"+"Space"+"/"+"2"/*URLProtocolVersion*/+"/"+Integer.toString(Server.User.UserID);
+		URL1 = "http://"+URL1+"/"+"Space"+"/"+"2"/*URLProtocolVersion*/+"/"+Long.toString(Server.User.UserID);
 		String URL2 = "Functionality"+"/"+"Components.dat";
 		//. add command parameters
 		URL2 = URL2+"?"+"1"/*command version*/+","+Integer.toString(idType)+","+Long.toString(idComponent);
